Componentize prefetched activity item summaries
