+++ components/Lists/ActivityList/InlineExpandedContent/ExpandedStateChannelCloseContent.js
