+++ components/Lists/ActivityList/InlineExpandedContent/index.js
@@ -1,0 +1,4 @@
+export { default as ExpandedPoCReceiptContent } from './ExpandedPoCReceiptContent'
+export { default as ExpandedPaymentContent } from './ExpandedPaymentContent'
+export { default as ExpandedStateChannelCloseContent } from './ExpandedStateChannelCloseContent'
+export { default as ExpandedRewardContent } from './ExpandedRewardContent'
