+++ components/Lists/ActivityList/PrefetchedSummaries/index.js
@@ -1,0 +1,4 @@
+export { default as PaymentV1Summary } from './PaymentV1Summary'
+export { default as PaymentV2Summary } from './PaymentV2Summary'
+export { default as PoCReceiptSummary } from './PoCReceiptSummary'
+export { default as RewardSummary } from './RewardSummary'
